Now the co-occurrence graph built has edge weights
HCSw call has been updated to use twice the mean
of all weights as the multiplier threshold,
as mirroring the original HCS usage of threshold of 2
when the unitary weights are used.
This way the dimension should matter less
(i.e. a graph with all weights equal 10 will be clustered
the same as a graph with all weights equal 1).